+++ scripts/archiso-base-211205.sh
@@ -1,0 +1,25 @@
+cat >> packages.x86_64 << /cat
+dhcpcd
+efibootmgr
+grub
+/cat
+sort -uo packages.x86_64 packages.x86_64
+
+ln -sv /usr/share/zoneinfo/Asia/Tokyo airootfs/etc/localtime
+
+cat > airootfs/etc/vconsole.conf << /cat
+KEYMAP=jp106
+/cat
+
+cat > airootfs/etc/hostname << /cat
+earth
+/cat
+
+ln -sv /usr/lib/systemd/system/dhcpcd.service /etc/systemd/system/multi-user.target.wants/
+
+mkdir airootfs/etc/sddm.conf.d
+cat > airootfs/etc/sddm.conf.d/autologin.conf << /cat
+[Autologin]
+User=user
+Session=plasma.desktop
+/cat